feat: implement net.fetch (#36733)
