+++ lib/browser/api/session.ts
@@ -1,4 +1,9 @@
-const { fromPartition, fromPath } = process._linkedBinding('electron_browser_session');
+import { fetchWithSession } from '@electron/internal/browser/api/net-fetch';
+const { fromPartition, fromPath, Session } = process._linkedBinding('electron_browser_session');
+
+Session.prototype.fetch = function (input: RequestInfo, init?: RequestInit) {
+  return fetchWithSession(input, init, this);
+};
 
 export default {
   fromPartition,
