+++ shell/browser/api/electron_api_session.cc
@@ -1228,10 +1228,16 @@ absl::optional<gin::Handle<Session>> Session::FromPath(
   return CreateFrom(isolate, browser_context);
 }
 
-gin::ObjectTemplateBuilder Session::GetObjectTemplateBuilder(
-    v8::Isolate* isolate) {
-  return gin_helper::EventEmitterMixin<Session>::GetObjectTemplateBuilder(
-             isolate)
+// static
+gin::Handle<Session> Session::New() {
+  gin_helper::ErrorThrower(JavascriptEnvironment::GetIsolate())
+      .ThrowError("Session objects cannot be created with 'new'");
+  return gin::Handle<Session>();
+}
+
+void Session::FillObjectTemplate(v8::Isolate* isolate,
+                                 v8::Local<v8::ObjectTemplate> templ) {
+  gin::ObjectTemplateBuilder(isolate, "Session", templ)
       .SetMethod("resolveProxy", &Session::ResolveProxy)
       .SetMethod("getCacheSize", &Session::GetCacheSize)
       .SetMethod("clearCache", &Session::ClearCache)
@@ -1301,7 +1307,8 @@ gin::ObjectTemplateBuilder Session::GetObjectTemplateBuilder(
       .SetProperty("protocol", &Session::Protocol)
       .SetProperty("serviceWorkers", &Session::ServiceWorkerContext)
       .SetProperty("webRequest", &Session::WebRequest)
-      .SetProperty("storagePath", &Session::GetPath);
+      .SetProperty("storagePath", &Session::GetPath)
+      .Build();
 }
 
 const char* Session::GetTypeName() {
@@ -1349,6 +1356,7 @@ void Initialize(v8::Local<v8::Object> exports,
                 void* priv) {
   v8::Isolate* isolate = context->GetIsolate();
   gin_helper::Dictionary dict(isolate, exports);
+  dict.Set("Session", Session::GetConstructor(context));
   dict.SetMethod("fromPartition", &FromPartition);
   dict.SetMethod("fromPath", &FromPath);
 }
